feat: moves Headline component out to common components.
