+++ src/routes/about.svelte
@@ -1,12 +1,8 @@
 <script>
   import Divider from "../components/common/Divider.svelte";
 
-  import Headline from "../components/home/Headline.svelte";
+  import Headline from "../components/common/Headline.svelte";
   import About from "../components/common/About.svelte";
-  import Work from "../components/home/Work.svelte";
-  import Skills from "../components/home/Skills.svelte";
-  import Instagram from "../components/home/Instagram.svelte";
-  import Footer from "../components/Footer.svelte";
 </script>
 
 <style>
@@ -21,11 +17,7 @@
 </svelte:head>
 
 <div>
+  <Headline title="About Me" />
   <About reverse />
   <Divider />
-  <Work />
-  <Divider />
-  <Skills />
-  <Divider />
-  <Instagram />
 </div>
