+++ src/routes/index.svelte
@@ -1,7 +1,7 @@
 <script>
   import Divider from "../components/common/Divider.svelte";
 
-  import Headline from "../components/home/Headline.svelte";
+  import Headline from "../components/common/Headline.svelte";
   import About from "../components/common/About.svelte";
   import Work from "../components/home/Work.svelte";
   import Skills from "../components/home/Skills.svelte";
@@ -21,7 +21,10 @@
 </svelte:head>
 
 <div>
-  <Headline />
+  <Headline
+    image
+    title="Hello, I’m Kevin, a React Native developer based in Stockholm,
+    Sweden." />
   <Divider />
   <About title="About Me" />
   <Divider />
